+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project does NOT adhere to [Semantic Versioning](https://semver.org/spe
 * infomaniak.py: add `insecure=False, no_proxy=False, timeout=8` to all functions
 * jitsi.py: no more hardcoded `insecure` and evaluate `no_proxy` in `get_data()`
 * librenms.py: Introduce caching of the result sets.
+* net.py: add `insecure=False, no_proxy=False, timeout=3` to `get_public_ip()`